+++ Makefile
@@ -6,7 +6,7 @@
 #    By: bperron <bperron@student.42.fr>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/06/26 13:16:40 by bperron           #+#    #+#              #
-#    Updated: 2022/09/28 13:17:14 by bperron          ###   ########.fr        #
+#    Updated: 2022/10/04 11:37:57 by bperron          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,7 +20,7 @@ LIBFT =  @$(MAKE) -sC ./libft
 LIBFTA = ./libft/libft.a
 
 HEADERS = include/minishell.h
-SRCS = src/minishell.c src/signal.c src/utils.c src/parsing.c src/built_in.c src/find_cmd.c src/cd_pwd.c src/variables.c src/exit.c src/export_utils.c src/echo.c src/check_pipe.c
+SRCS = src/minishell.c src/signal.c src/utils.c src/parsing.c src/built_in.c src/find_cmd.c src/cd_pwd.c src/variables.c src/exit.c src/export_utils.c src/echo.c src/check_pipe.c src/change_var.c
 OBJS = $(SRCS:.c=.o)
 
 all: $(NAME)
